Add JWT authentication flow
Add the ability for users to authenticate using JWT.  A JWT will be
created, signed, and exchanged for a bearer token which can be used to
authenticate with the Adobe Analytics API V2.0.

Authentication using a JWT requires that a JWT integration with proper
permissions be setup in the Adobe I/O Console.  Instructions on how to
complete this can be found in the Adobe API documentation:

https://www.adobe.io/apis/experiencecloud/analytics/docs.html#!AdobeDocs/analytics-2.0-apis/master/jwt.md

Once the integration has been created, users wishing to authenticate
with JWT will need to download the config file and ensure that they have
access to the RSA private key file which corresponds to the public key
certificate uploaded during the integration creation.

The file paths to the config file and the private key file can be set in
the environment variables AW_JWT_CONFIG_FILE and AW_JWT_PRIVATE_KEY_FILE
respectively.  The auth flow will then be able to use these environment
variable values to load the correct configuration settings and create
the JWT.

Authenticating using the JWT integration can be accomplished using the
`auth_jwt()` method.